add WithActiveTab HOC and connect it
