+++ src/components/tabs/tabs.jsx
@@ -2,7 +2,12 @@ import React from "react";
 import PropTypes from "prop-types";
 import {TabType} from "../../consts";
 
-const Tabs = ({onTabClick, activeTab}) => {
+import OverviewTab from "../overview-tab/overview-tab.jsx";
+import DetailsTab from "../details-tab/details-tab.jsx";
+import ReviewsTab from "../reviews-tab/reviews-tab.jsx";
+
+
+const Tabs = ({onTabClick, activeTab, movie}) => {
   const tabValues = Object.values(TabType);
 
   const tabClickHandler = (value) => {
@@ -12,7 +17,27 @@ const Tabs = ({onTabClick, activeTab}) => {
     };
   };
 
-  const renderTab = (value, i) => {
+  const renderTab = (currentTab) => {
+
+    const TabPages = {
+      Overview:
+        <OverviewTab
+          movie = {movie}
+        />,
+      Details:
+        <DetailsTab
+          movie={movie}
+        />,
+      Reviews:
+        <ReviewsTab
+          reviews={movie.reviews}
+        />
+    };
+
+    return TabPages[currentTab];
+  };
+
+  const renderTabLink = (value, i) => {
     const tabClassName = `movie-nav__item ${value === activeTab ? `movie-nav__item--active` : ``}`;
     const navKey = value + i;
 
@@ -25,12 +50,18 @@ const Tabs = ({onTabClick, activeTab}) => {
     );
   };
 
-  const renderTabsList = () => tabValues.map(renderTab);
+  const renderTabsList = () => tabValues.map(renderTabLink);
 
   return (
-    <ul className="movie-nav__list">
-      {renderTabsList()}
-    </ul>
+    <div className="movie-card__desc">
+
+      <nav className="movie-nav movie-card__nav">
+        <ul className="movie-nav__list">
+          {renderTabsList()}
+        </ul>
+      </nav>
+      {renderTab(activeTab)}
+    </div>
   );
 };
